requirements.txt
git+https://github.com/ntasfi/PyGame-Learning-Environment.git
git+https://github.com/simoninithomas/gym-games
huggingface_hub
imageio-ffmpeg
pyyaml==6.0
matplotlib
torch
numpy==1.23.5
